+++ packages/gamut/src/GridForm/types.ts
@@ -1,4 +1,4 @@
-import { ValidationOptions } from 'react-hook-form';
+import { FormContextValues, ValidationOptions } from 'react-hook-form';
 
 import { ColumnSizes, ResponsiveProperty } from '../Layout';
 
@@ -11,12 +11,30 @@ export type GridFormCheckboxField = BaseFormField & {
   description: string;
   defaultValue?: boolean;
   label?: string;
+  validation?: Pick<ValidationOptions, 'required'>;
   type: 'checkbox';
+};
+
+export type GridFormCustomFieldProps = {
+  className: string;
+  error?: string;
+  field: GridFormCustomField;
+  register: FormContextValues['register'];
+  setValue: (value: any) => void;
+};
+
+export type GridFormCustomField = BaseFormField & {
+  defaultValue?: any;
+  label?: string;
+  render: (props: GridFormCustomFieldProps) => React.ReactNode;
+  validation?: ValidationOptions;
+  type: 'custom';
 };
 
 export type GridFormTextField = BaseFormField & {
   defaultValue?: string;
   label: string;
+  placeholder?: string;
   validation?: ValidationOptions;
   type: 'text' | 'email';
 };
@@ -45,6 +63,7 @@ export type GridFormTextAreaField = BaseFormField & {
 
 export type GridFormField =
   | GridFormCheckboxField
+  | GridFormCustomField
   | GridFormTextField
   | GridFormSelectField
   | GridFormFileField