+++ packages/styleguide/stories/Brand/Quote.stories.tsx
@@ -3,15 +3,16 @@ import { text } from '@storybook/addon-knobs';
 import React from 'react';
 
 import {
+  decoratedStories,
   decoratedStory,
   StoryDescription,
   StoryStatus,
   StoryTemplate,
 } from '../Templating';
 
-export default decoratedStory('Brand', Quote);
+export default decoratedStories('Brand', Quote);
 
-export const quote = () => (
+export const quote = decoratedStory(() => (
   <StoryTemplate status={StoryStatus.Ready}>
     {theme => (
       <>
@@ -29,12 +30,4 @@ export const quote = () => (
       </>
     )}
   </StoryTemplate>
-);
-
-quote.story = {
-  parameters: {
-    knobs: {
-      escapeHTML: false,
-    },
-  },
-};
+));